+++ CHANGELOG.md
@@ -13,5 +13,6 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `zsm create` command which allows to create snapshots of a single
   specified or all availalbe file systems.
 * `zsm clean` command which allows to remove obsolete snapshots.
+* `zsm list` command which allows to list all snapshots managed by zsm.
 
 [Unreleased]: https://github.com/fhofherr/zsm